chore: move error definition into `-private`
I don't think it makes sense to consider this part of the public,
importable API. We can always choose to make it "public" later on if
users seem to have a need for it
