+++ addon/-private/errors.js
